+++ src/main/kotlin/org/polyfrost/evergreenhud/client/hud/ResourcePackHud.kt
@@ -20,7 +20,6 @@ import org.polyfrost.compose.composables.PolyMcText
 import org.polyfrost.compose.composables.PolyModifier
 import org.polyfrost.compose.composables.PolyRow
 import org.polyfrost.compose.composables.PolyText
-import org.polyfrost.compose.composables.background
 import org.polyfrost.compose.composables.padding
 import org.polyfrost.compose.composables.size
 import org.polyfrost.compose.mc.McFontQueue
@@ -178,6 +177,8 @@ class ResourcePackHud : Hud(
 
     override fun defaultPosition(): Pair<Float, Float> = 0f to 0f
 
+    override fun canMergeBackground(): Boolean = true
+
     override fun updateFrequency(): Long = 1.seconds.inWholeNanoseconds
 
     override fun setup() {
@@ -215,12 +216,7 @@ class ResourcePackHud : Hud(
 
         val lines = wrap(packDescription.value, DESCRIPTION_WIDTH * scale, measure)
 
-        var modifier = PolyModifier.padding(padLeft, padTop, padRight, padBottom)
-        if (showBackground) {
-            modifier = PolyModifier
-                .background(PolyColor(bgColor, bgChroma, bgChromaSpeed), bgRadius)
-                .padding(padLeft, padTop, padRight, padBottom)
-        }
+        val modifier = hudBackground().padding(padLeft, padTop, padRight, padBottom)
 
         PolyBox(modifier = modifier) {
             PolyRow(gap = GAP * scale) {